Fix agenda date key helper
